fix additional gii errors

+++ cortex/brainctm.py
@@ -149,7 +149,7 @@ class Hemi(object):
         self.tf = tempfile.NamedTemporaryFile()
         self.ctm = CTMfile(self.tf.name, "w")
 
-        self.ctm.setMesh(pts, polys, norms=norms)
+        self.ctm.setMesh(pts.astype(np.float32), polys, norms=norms)
 
         self.pts = pts
         self.polys = polys
